pyproject.toml
[project]
name = "plantvillage"
version = "0.1.0"
description = "Deep Learning solution for detecting diseases in plants."
authors = [
    {name = "Rohan Khaire"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch (>=2.6.0,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "torchvision (>=0.21.0,<0.22.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "matplotlib (>=3.10.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
